Visual C++ 2010 Express support
fixes ValveSoftware/halflife#1338

Changes:
- added project files for Visual C++ 2010 Express
- resolved function calls that would otherwise be ambiguous now
- solved ARRAYSIZE macro and HSPRITE type collisions with Windows.h and
  winsock.h includes by guarding the includes

TODO:
We should think about if we actually should instead link to a .h that
defines / redirects the required macros, so that platform differences
would be limited to a few .cpp files (meaning not including windows.h in a
huge number of files as it is now), see the issue above for details.

+++ cl_dll/hud_spectator.cpp
@@ -1428,7 +1428,7 @@ void CHudSpectator::DrawOverviewLayer()
 	if ( hasMapImage)
 	{
 		i = m_MapSprite->numframes / (4*3);
-		i = sqrt(i);
+		i = sqrt((float)i);
 		xTiles = i*4;
 		yTiles = i*3;
 	}
